Another big round of changes to fix gcc warnings / ensure standards conformance.
Changes include 
- explicit casting between types
- removing or commenting out unused variables / functions
- correct using of "const" modifier
- replacing use of finite with isfinite where available (this is in C99, but not supported by MSVC so still need complicated set of macros to check)
- rename local variables which shadow other global or local variables / symbols


svn path=/trunk/; revision=11973

+++ drivers/xfig.c
@@ -166,7 +166,7 @@ plD_init_xfig( PLStream *pls )
     stcmap1( pls );
 
     dev->bufflen = 2 * BSIZE;
-    dev->buffptr = (int *) malloc( sizeof ( int ) * dev->bufflen );
+    dev->buffptr = (int *) malloc( sizeof ( int ) * (size_t) ( dev->bufflen ) );
     if ( dev->buffptr == NULL )
         plexit( "plD_init_xfig: Out of memory!" );
 }
@@ -265,7 +265,7 @@ plD_line_xfig( PLStream *pls, short x1a, short y1a, short x2a, short y2a )
         {
             dev->bufflen += 2 * BSIZE;
             tempptr       = (int *)
-                            realloc( (void *) dev->buffptr, dev->bufflen * sizeof ( int ) );
+                            realloc( (void *) dev->buffptr, (size_t) ( dev->bufflen ) * sizeof ( int ) );
             if ( tempptr == NULL )
             {
                 free( (void *) dev->buffptr );
